Update linux-postmarketos-mainline to 4.15_rc6
cgit doesn't generate proper patches for binary files (one was added in
v4.15-rc1 and removed in v4.15-rc5 because it also introduced a new
build-time dependency which is now solved. For 4.16 the "old" source
lines can be re-added but until then a full diff skipping v4.15-rc1
is needed.